auth added

+++ src/main/java/io/swagger/sample/resource/PetStoreResource.java
@@ -19,11 +19,11 @@ package io.swagger.sample.resource;
 import io.swagger.annotations.*;
 import io.swagger.sample.data.PetData;
 import io.swagger.sample.data.StoreData;
-import io.swagger.sample.exception.BadRequestException;
+import io.swagger.sample.exception.ApiException;
 import io.swagger.sample.model.AbstractApiResponse;
 import io.swagger.sample.model.Inventory;
 import io.swagger.sample.model.Order;
-import io.swagger.sample.exception.NotFoundException;
+import io.swagger.sample.util.AuthFilter;
 import io.swagger.util.Json;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,8 +33,8 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.*;
 
 @Path("/store")
-@Api(value="/store", tags = "store")
-@Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
+@Api(value = "/store", tags = "store")
+@Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
 public class PetStoreResource {
   static StoreData storeData = new StoreData();
   static PetData petData = new PetData();
@@ -43,53 +43,42 @@ public class PetStoreResource {
 
   @GET
   @Path("/inventory")
-  @Produces({MediaType.APPLICATION_JSON})
-  @ApiOperation(value = "Returns pet inventories by status",
-    notes = "Returns a map of status codes to quantities",
-    response = Inventory.class,
-    authorizations = @Authorization(value = "api_key")
-  )
-  public java.util.Map<String, Integer> getInventory() {
+  @Produces({ MediaType.APPLICATION_JSON })
+  @ApiOperation(value = "Returns pet inventories by status", notes = "Returns a map of status codes to quantities", response = Inventory.class)
+  @ApiResponses(value = { @ApiResponse(code = 400, message = "Invalid ID supplied"),
+      @ApiResponse(code = 401, message = "Must have admin permissions to access this endpoint") })
+  public java.util.Map<String, Integer> getInventory(
+      @ApiParam(value = "Admin only", required = true) @HeaderParam("token") String token) throws ApiException {
+    AuthFilter.assertAdmin(token);
     return petData.getInventoryByStatus();
   }
 
   @GET
   @Path("/order/{orderId}")
-  @ApiOperation(value = "Find purchase order by ID",
-    notes = "For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions",
-    response = Order.class)
-  @ApiResponses(value = { @ApiResponse(code = 400, message = "Invalid ID supplied"),
+  @ApiOperation(value = "Find purchase order by ID", notes = "For valid response try integer IDs with value >= 1 and <= 10", response = Order.class)
+  @ApiResponses(value = { @ApiResponse(code = 401, message = "Must have admin permissions to access this endpoint"),
       @ApiResponse(code = 404, message = "Order not found") })
-  public Response getOrderById(
-      @ApiParam(value = "ID of pet that needs to be fetched", allowableValues = "range[1,10]", required = true)
-      @PathParam("orderId") Long orderId)
-      throws NotFoundException {
-    LOGGER.debug("getOrderById {}", orderId);
+  public Response getOrderById(@ApiParam(value = "Admin only", required = true) @HeaderParam("token") String token,
+      @ApiParam(value = "ID of pet that needs to be fetched", required = true) @PathParam("orderId") Long orderId)
+      throws ApiException {
+    AuthFilter.assertAdmin(token);
     Order order = storeData.findOrderById(orderId);
     if (null != order) {
       return Response.ok().entity(order).build();
-    } else {
-      throw new NotFoundException(404, "Order not found");
     }
+    throw new ApiException(404, "Order not found");
   }
 
   @POST
   @Path("/order")
   @ApiOperation(value = "Place an order for a pet", consumes = "application/json", response = Order.class)
-  @ApiResponses({ @ApiResponse(code = 400, message = "Invalid Order") })
-  public Order placeOrder(
-      @ApiParam(value = "order placed for purchasing the pet",
-        required = true) Order order) throws BadRequestException {
+  @ApiResponses({ @ApiResponse(code = 401, message = "Must have user permissions to access this endpoint"),
+      @ApiResponse(code = 400, message = "Wrong Order structure") })
+  public Order placeOrder(@ApiParam(value = "User only", required = true) @HeaderParam("token") String token,
+      @ApiParam(value = "order placed for purchasing the pet", required = true) Order order) throws ApiException {
+    AuthFilter.assertUser(token);
     if (order == null) {
-      throw new BadRequestException(400, "No data");
-    }
-    try {
-      LOGGER.info("placeOrder ID {} STATUS {}", order.getId(), order.getStatus());
-      if (LOGGER.isDebugEnabled()) {
-        LOGGER.debug("placeOrder {}", Json.mapper().writeValueAsString(order));
-      }
-    } catch (Throwable e) {
-      e.printStackTrace();
+      throw new ApiException(400, "Wrong Order structure");
     }
     storeData.placeOrder(order);
     return storeData.placeOrder(order);
@@ -97,18 +86,17 @@ public class PetStoreResource {
 
   @DELETE
   @Path("/order/{orderId}")
-  @ApiOperation(value = "Delete purchase order by ID",
-    notes = "For valid response try integer IDs with positive integer value. Negative or non-integer values will generate API errors")
-  @ApiResponses(value = { @ApiResponse(code = 400, message = "Invalid ID supplied"),
+  @ApiOperation(value = "Delete purchase order by ID")
+  @ApiResponses(value = { @ApiResponse(code = 401, message = "Must have admin permissions to access this endpoint"),
       @ApiResponse(code = 404, message = "Order not found") })
-  public Response deleteOrder(
-      @ApiParam(value = "ID of the order that needs to be deleted", allowableValues = "range[1,infinity]", required = true)
-      @PathParam("orderId") Long orderId) {
-    LOGGER.debug("deleteOrder {}", orderId);
+  public Response deleteOrder(@ApiParam(value = "Admin only", required = true) @HeaderParam("token") String token,
+      @ApiParam(value = "ID of the order that needs to be deleted", required = true) @PathParam("orderId") Long orderId)
+      throws ApiException {
+    AuthFilter.assertAdmin(token);
     if (storeData.deleteOrder(orderId)) {
       return Response.ok().entity(new AbstractApiResponse(String.valueOf(orderId))).build();
     } else {
-      return Response.status(Response.Status.NOT_FOUND).entity(new AbstractApiResponse(String.valueOf("Order Not Found"))).build();
+      throw new ApiException(404, "Order not found");
     }
   }
 }
